final changes with badges added
