--- application/helper_functions/representations.py
@@ -1,7 +1,0 @@
-from dicttoxml import dicttoxml
-
-
-def text_xml(data, code):
-    '''changes default datatype pf the response to text/xml'''
-    resp = (dicttoxml(data), code)
-    return resp
