--- my_application/helper_functions/get_allele.py
@@ -1,9 +1,0 @@
-def get_allele(vr_response):
-    ''' iterates over response from variant recoder to get list of alleles present'''
-    keys = [] 
-       
-    for dictionary in vr_response:
-        for key in dictionary:
-            keys.append(key)
-                                               
-        return keys
