--- my_application/helper_functions/get_xml.py
@@ -1,7 +1,0 @@
-from dicttoxml import dicttoxml
-
-
-def text_xml(data):
-    '''changes default datatype pf the response to text/xml'''
-    resp = dicttoxml(data)
-    return resp
