+++ pytest.ini
@@ -1,6 +1,5 @@
 [pytest]
 minversion = 7.0
-#addopts = --cov=src
-pythonpath = my_application tests
+pythonpath = variant_annotator tests
 testpaths =
     tests
