+++ variant_annotator/get_functions/get_VEP.py
@@ -1,8 +1,8 @@
 from helper_functions import get_requests
 
-def VariantEPredictor(species, hgvs_notation): #allows options to be optional
+def VariantEPredictor(species, hgvs_notation):
         '''Determines the effect of the variant on gene, transcript, protein and regulaotry regions'''
-
+        # Specify the url to request data from
         url = f"https://rest.ensembl.org/vep/{species}/hgvs/{hgvs_notation}/"
         
-        return get_requests.request_vepdata(url) 
+        return get_requests.request_vepdata(url)
